Batch Services refactoring
