+++ src/batches/batchDetails.tsx
@@ -23,9 +23,6 @@ import ViewLogs from '../utils/viewLogs';
 import DeleteClusterIcon from '../../style/icons/delete_cluster_icon.svg';
 import { ClipLoader } from 'react-spinners';
 import {
-  API_HEADER_BEARER,
-  API_HEADER_CONTENT_TYPE,
-  BASE_URL,
   DATAPROC_CLUSTER_KEY,
   DATAPROC_CLUSTER_LABEL,
   BATCH_FIELDS_EXCLUDED,
@@ -45,24 +42,19 @@ import {
 import {
   BatchTypeValue,
   IBatchInfoResponse,
-  authApi,
   batchDetailsOptionalDisplay,
   convertToDCUHours,
   convertToGBMonths,
   elapsedTime,
   jobTimeFormat,
-  statusMessageBatch,
-  toastifyCustomStyle,
-  loggedFetch
+  statusMessageBatch
 } from '../utils/utils';
 import DeletePopup from '../utils/deletePopup';
-import { toast } from 'react-toastify';
 import 'react-toastify/dist/ReactToastify.css';
-import { deleteBatchAPI } from './batchService';
+import { deleteBatchAPI, getBatchDetailsInfo } from './batchService';
 import { statusDisplay } from '../utils/statusDisplay';
 import PollingTimer from '../utils/pollingTimer';
 import CreateBatch from './createBatch';
-import { DataprocLoggingService, LOG_LEVEL } from '../utils/loggingService';
 
 const iconLeftArrow = new LabIcon({
   name: 'launcher:left-arrow-icon',
@@ -86,7 +78,7 @@ type BatchDetailsProps = {
 function BatchDetails({
   batchSelected,
   setDetailedBatchView,
-  setCreateBatchView,
+  setCreateBatchView
 }: BatchDetailsProps) {
   const [batchInfoResponse, setBatchInfoResponse] = useState({
     uuid: '',
@@ -175,111 +167,16 @@ function BatchDetails({
       pollingBatchDetails(getBatchDetails, true);
     };
   }, []);
-  interface IBatchDetailsResponse {
-    error: {
-      code: number;
-      message: string;
-    };
-    uuid: '';
-    state: '';
-    createTime: '';
-    runtimeInfo: {
-      endpoints: {};
-      approximateUsage: { milliDcuSeconds: ''; shuffleStorageGbSeconds: '' };
-    };
-    creator: '';
-    runtimeConfig: {
-      version: '';
-      containerImage: '';
-      properties: {
-        'spark:spark.executor.instances': '';
-        'spark:spark.driver.cores': '';
-        'spark:spark.driver.memory': '';
-        'spark:spark.executor.cores': '';
-        'spark:spark.executor.memory': '';
-        'spark:spark.dynamicAllocation.executorAllocationRatio': '';
-        'spark:spark.app.name': '';
-      };
-    };
-    sparkBatch: {
-      mainJarFileUri: '';
-      mainClass: '';
-      jarFileUris: '';
-    };
-    pysparkBatch: {
-      mainPythonFileUri: '';
-    };
-    sparkRBatch: {
-      mainRFileUri: '';
-    };
-    sparkSqlBatch: {
-      queryFileUri: '';
-    };
-    environmentConfig: {
-      executionConfig: {
-        serviceAccount: '';
-        subnetworkUri: '';
-        networkTags: [];
-        kmsKey: '';
-      };
-      peripheralsConfig: {
-        metastoreService: '';
-        sparkHistoryServerConfig: {
-          dataprocCluster: '';
-        };
-      };
-    };
-    stateHistory: [{ state: ''; stateStartTime: '' }];
-    stateTime: '';
-    labels: {};
-  }
 
   const getBatchDetails = async () => {
-    const credentials = await authApi();
-    if (credentials) {
-      setRegionName(credentials.region_id || '');
-      setProjectName(credentials.project_id || '');
-      loggedFetch(
-        `${BASE_URL}/projects/${credentials.project_id}/locations/${credentials.region_id}/batches/${batchSelected}`,
-        {
-          method: 'GET',
-          headers: {
-            'Content-Type': API_HEADER_CONTENT_TYPE,
-            Authorization: API_HEADER_BEARER + credentials.access_token
-          }
-        }
-      )
-        .then((response: Response) => {
-          response
-            .json()
-            .then((responseResult: IBatchDetailsResponse) => {
-              setBatchInfoResponse(responseResult);
-              if (responseResult.labels) {
-                const labelValue = Object.entries(responseResult.labels).map(
-                  ([key, value]) => `${key}:${value}`
-                );
-                setLabelDetail(labelValue);
-              }
-              setIsLoading(false);
-              if (responseResult?.error?.code) {
-                toast.error(responseResult?.error?.message, toastifyCustomStyle);
-              }
-            })
-            .catch((e: Error) => {
-              console.log(e);
-              setIsLoading(false);
-            });
-        })
-        .catch((err: Error) => {
-          setIsLoading(false);
-          DataprocLoggingService.log('Error in getting Batch details', LOG_LEVEL.ERROR);
-          console.error('Error in getting Batch details', err);
-          toast.error(
-            `Failed to fetch batch details ${batchSelected}`,
-            toastifyCustomStyle
-          );
-        });
-    }
+    await getBatchDetailsInfo(
+      setRegionName,
+      setProjectName,
+      batchSelected,
+      setBatchInfoResponse,
+      setLabelDetail,
+      setIsLoading
+    );
   };
 
   const statusMsg = statusMessageBatch(batchInfoResponse);
@@ -364,9 +261,7 @@ function BatchDetails({
 
       {!createBatch && batchInfoResponse.uuid !== '' && (
         <div className="scroll">
-          <div
-            className='scroll-fix-header cluster-details-header'
-          >
+          <div className="scroll-fix-header cluster-details-header">
             <div
               role="button"
               className="back-arrow-icon"
@@ -374,7 +269,7 @@ function BatchDetails({
             >
               <iconLeftArrow.react
                 tag="div"
-                className='icon-white logo-alignment-style'
+                className="icon-white logo-alignment-style"
               />
             </div>
             <div className="cluster-details-title">{batchSelected}</div>
